Fix integration tests
The integration tests failed because an element selector found two links
instead of one, after adding a link element wrapping the media card
image.

This happened because the element selector was looking for a link with
an "accessible name"  matching the media title. Previously this was only
the media card title, which was a link with the title in it, so
the accessible name matched. But with the new link around the image,
that link also got the title of the media as an accessible name because
the title was used as an alt for the image.

This has been resolved by getting the element by its title instead of its
role, since the title attribute is explicit instead of implicit and is
only set on the desired title link element.